Improve evaluation of abstract conditionals (#2503)
Summary:
Release notes: none

Currently, when don't really fully deal with abstract conditions fully throughout the Prepack codebase and we usually simply leak/emit a temporal to get around them. So this PR aims at tackling many of the code-sites where they are prevalent by using the internal React bundle to find most of them and address them – with the goal of improving evaluation in cases where we run into conditionals.

With this PR, we now try and resolve different parts of the condition, including conditionals such as `||` and `&&`. This allows us to evaluate potentially far more than before when it comes to product code where we inhabit many deep conditionals. These changes also impact `Object.assign` as it triggered an invariant with `getSnapshot` as it never expected a conditional there. This should also fix https://github.com/facebook/prepack/issues/2323.

I've added some tests to show this but here is an example of the output before and after:

```js
// Input
function fn(x, b) {
  var a = x ? b : { a: 2 };
  return a.a;
}

// Before
  var _2 = function (x, b) {
    var _3 = {};
    _3.a = 2;
    var _$0 = (x ? b : _3).a;
    return _$0;
  };

// After
  var _2 = function (x, b) {
    if (x) {
      var _$0 = b.a;
    }
    return x ? _$0 : 2;
  };
```

For conditional `Object.assign`, the output looks like this (note: we need materializing rather than leaking to better improve the output):
```js
// Input
function fn(x) {
  var a = x ? { a: 1 } : { a: 2 };

  return Object.assign({}, a, { b: 1 });
}

// Before
var _2 = function (x) {
  var _$0 = {
    a: 1
  };
  var _$1 = {
    a: 2
  };
  ({}).a = 1;
  ({}).a = 2;
  var _$2 = {
    b: 1
  };
  var _9 = {};

  var _$3 = _$9(_9, x ? _$0 : _$1, _$2);

  return _9;
};

// After
  var _2 = function (x) {
    return {
      a: x ? 1 : 2,
      b: 1
    };
  };
```

I also took the time to apply the same small changes to the existing code in `CallExpression` so the logic there could also handle `&&` and `||` cases too. Including is a React test that shows that we can now inline a component that we previously weren't able to do.
Pull Request resolved: https://github.com/facebook/prepack/pull/2503

Differential Revision: D9616500

Pulled By: trueadm

fbshipit-source-id: 3888a62da330c64b0395723f8764c3590adc8491