Fix problem where a bind call can use a variable before its definition (#2558)
Summary:
Release note: none

A large internal test case came up with a situation where an optimized function is rewritten with a version that early binds a yet to be declared value to one of its parameters.

This should probably never happen, so the if condition guarding this code path has been strengthened to prevent this.
Pull Request resolved: https://github.com/facebook/prepack/pull/2558

Reviewed By: trueadm

Differential Revision: D9967064

Pulled By: hermanventer

fbshipit-source-id: bf56872ea42ee8faefc347a6393cdcd9b3c518e9