Adds createdAbstracts Set to effects/realm for tracking (#2593)
Summary:
Release notes: adds `createdAbstracts` `Set` to both `Effects` and `Realm`.

For a feature I'm working on, where we can optionally inline function calls depending on a bunch of pre-defined heuristics, I've run into the need for a feature for tracking abstract values. When we use `evaluateForEffects` we currently get quite a bit of information of what occurred, nicely packaged up in effects. One of those is `createdObjects`, which is super useful for many cases, but I believe we also need, is a way of knowing what `AbstractValue`s were created in effects too. This PR adds that functionality.
Pull Request resolved: https://github.com/facebook/prepack/pull/2593

Differential Revision: D10276125

Pulled By: trueadm

fbshipit-source-id: 299b0f91d93b1491987b10e769ca77f17e64716d