Split mechanism and policy in the leaking visitor (#2594)
Summary:
The leaking, havocing, materialization and alias tracking implementation consists of two parts. The first part helps compute values reachable from other values with various constraints. The second part is the one that actually uses these reachable values and performs the action needed. This PR splits the first (mechanism) part from the second (policy) part.

Reachability can be used by invoking the following function:

```js
computeReachableObjectsAndBindings(
    realm: Realm,
    rootValue: Value,
    filterValue: Value => boolean,
    readOnly?: boolean
  ): [Set<ObjectValue>, Set<Binding>]
```
...to compute all possible values that `rootValue` could evaluate to. It returns a tuple consisting of a set of reachable objects and a set of reachable bindings.

`filterValue` limits the values that we traverse. For leaking, `filterValue` checks that a value has not already been leaked (since interactions with leaked values cause leaking), if a value is certainly not an object, if it is frozen, in scope etc. You could use this to find particular types of reachable objects.

If the `readOnly` flag is set, then only read bindings are returned.
Pull Request resolved: https://github.com/facebook/prepack/pull/2594

Differential Revision: D10488695

Pulled By: sb98052

fbshipit-source-id: c8f0a63148f4cccd6434ebe07bebfb38dcd5cf08